# src/main/resources/static/openapi/SupportAdditionalNeedsAPI.yml
openapi: 3.1.1
info:
  title: Support For Additional Needs API
  version: '0.13.7'
  description: Support For Additional Needs API
  contact:
    name: Learning and Work Progress team
    email: learningandworkprogress@digital.justice.gov.uk
servers:
  - url: https://support-for-additional-needs-api-dev.hmpps.service.justice.gov.uk
    description: Development
  - url: https://support-for-additional-needs-api-preprod.hmpps.service.justice.gov.uk
    description: Pre-Production
  - url: https://support-for-additional-needs-api.hmpps.service.justice.gov.uk
    description: Production
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: [ ]

#
# --------------------------------------------------------------------------------
#
paths:
...
  '/profile/{prisonNumber}/challenges':
    parameters:
      - $ref: "#/components/parameters/prisonNumberPathParameter"
    post:
      summary: Adds one or more challenges.
      description: |
        Adds one or more Challenges to the specified prisoner's list of Challenges.
                  
        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
      tags:
        - Challenge
      responses:
        '201':
          $ref: '#/components/responses/ChallengeList'
        '400':
          $ref: '#/components/responses/400Error'
        '409':
          $ref: '#/components/responses/409Error'
      operationId: post-challenges
      requestBody:
        $ref: '#/components/requestBodies/CreateChallengeList'
    get:
      summary: Retrieves the list of challenges.
      description: |

        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RO` (Read-Only)
      tags:
        - Challenge
      responses:
        '200':
          $ref: '#/components/responses/ChallengeList'
        '400':
          $ref: '#/components/responses/400Error'
      operationId: get-challenges
  '/profile/{prisonNumber}/challenges/{challengeReference}':
    parameters:
      - $ref: "#/components/parameters/prisonNumberPathParameter"
      - name: challengeReference
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: UUID of the specific challenge to update
    get:
      summary: Retrieves a challenge by reference.
      description: |

        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RO` (Read-Only)
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
      tags:
        - Challenge
      responses:
        '200':
          $ref: '#/components/responses/Challenge'
        '404':
          $ref: '#/components/responses/404Error'
      operationId: get-challenge
    put:
      summary: Updates a single challenge by UUID.
      description: |
        Updates a specific challenge associated with the given prison number.

        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)

      tags:
        - Challenge
      operationId: update-challenge
      requestBody:
        $ref: '#/components/requestBodies/UpdateChallenge'
      responses:
        '200':
          $ref: '#/components/responses/Challenge'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          description: Challenge not found
...
  '/profile/{prisonNumber}/support-strategies':
    parameters:
      - $ref: "#/components/parameters/prisonNumberPathParameter"
    post:
      summary: Adds one or more support strategies.
      description: |
        Adds one or more Support Strategies to the specified prisoner's list of Support Strategies.

        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
      tags:
        - Support Strategy
      responses:
        '201':
          $ref: '#/components/responses/SupportStrategyList'
        '400':
          $ref: '#/components/responses/400Error'
        '409':
          $ref: '#/components/responses/409Error'
      operationId: post-support-strategies
      requestBody:
        $ref: '#/components/requestBodies/CreateSupportStrategyList'

    get:
      summary: Retrieves the list of support strategies.
      description: |
        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RO` (Read-Only)
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
      tags:
        - Support Strategy
      responses:
        '200':
          $ref: '#/components/responses/SupportStrategyList'
        '400':
          $ref: '#/components/responses/400Error'
      operationId: get-support-strategies
  '/profile/{prisonNumber}/support-strategies/{supportStrategyReference}':
    parameters:
      - $ref: "#/components/parameters/prisonNumberPathParameter"
      - name: supportStrategyReference
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: UUID of the specific support strategy to update
    get:
      summary: Retrieves a support strategy.
      description: |
        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RO` (Read-Only)
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
      tags:
        - Support Strategy
      responses:
        '200':
          $ref: '#/components/responses/SupportStrategy'
        '400':
          $ref: '#/components/responses/404Error'
      operationId: get-support-strategy
    put:
      summary: Updates a single support strategy by UUID.
      description: |
        Updates a specific support strategy associated with the given prison number.

        **Role Requirements:**
        Access to this endpoint requires one of the following roles:
          - `ROLE_SUPPORT_ADDITIONAL_NEEDS__ELSP__RW` (Read-Write)
      tags:
        - Support Strategy
      operationId: update-support-strategy
      requestBody:
        $ref: '#/components/requestBodies/UpdateSupportStrategy'
      responses:
        '200':
          $ref: '#/components/responses/SupportStrategy'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          description: Support Strategy not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #
  # Request Body Definitions
  # --------------------------------------------------------------------------------
  requestBodies:
...
    CreateChallengeList:
      description: Request body to create challenges for a person.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateChallengesRequest'
    UpdateChallenge:
      description: Request body to update a challenge for a person.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateChallengeRequest'
...
    CreateSupportStrategyList:
      description: Request body to create support strategies for a person.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateSupportStrategiesRequest'
    UpdateSupportStrategy:
      description: Request body to update a support strategy for a person.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateSupportStrategyRequest'


  #
  # Schema and Enum Definitions
  # --------------------------------------------------------------------------------
  schemas:
    PrisonId:
      type: string
      description: The Prison identifier.
      example: BXI
      pattern: "^[A-Z]{3}$"
...
    ErrorResponse:
      title: ErrorResponse
      description: |
        A generic response body object describing errors in a web request, and can be used to communicate several
        different types of error condition such as (but not limited to) `400 BAD REQUEST`, `409 CONFLICT` etc.
      properties:
        status:
          description: The HTTP status code.
          type: integer
          example: 400
        errorCode:
          description: An optional application specific error code.
          type: string
        userMessage:
          description: An optional human readable description of the error.
          type: string
          example: 'No Plan found for prisoner [A1234BC]'
        developerMessage:
          description: An optional error message that may have more technical information for developers.
          type: string
        moreInfo:
          description: Optional more detailed information about the error.
          type: string
      required:
        - status
...
    AuditedResource:
      title: AuditedResource
      type: object
      properties:
        createdBy:
          type: string
          description: The DPS username of the person who created this resource.
          example: 'asmith_gen'
        createdByDisplayName:
          type: string
          description: The display name of the person who created this resource.
          example: 'Alex Smith'
        createdAt:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when this resource was created.
          example: '2023-06-19T09:39:44Z'
        createdAtPrison:
          type: string
          description: The identifier of the prison that the prisoner was resident at when this resource was created.
          example: 'BXI'
        updatedBy:
          type: string
          description: The DPS username of the person who last updated this resource.
          example: 'asmith_gen'
        updatedByDisplayName:
          type: string
          description: The display name of the person who last updated this resource.
          example: 'Alex Smith'
        updatedAt:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when this resource was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
        updatedAtPrison:
          type: string
          description: The identifier of the prison that the prisoner was resident at when this resource was updated.
          example: 'BXI'
      required:
        - createdBy
        - createdByDisplayName
        - createdAt
        - createdAtPrison
        - updatedBy
        - updatedByDisplayName
        - updatedAt
        - updatedAtPrison
...
    ChallengeListResponse:
      title: ChallengeListResponse
      description: A list of challenges for a prisoner.
      type: object
      properties:
        challenges:
          type: array
          description: A List containing zero or more Challenges.
          items:
            $ref: '#/components/schemas/ChallengeResponse'
      required:
        - challenges
...
    CreateChallengesRequest:
      title: ChallengesListRequest
      description: A request to create a list of challenges for a person.
      type: object
      properties:
        challenges:
          type: array
          description: A List containing zero or more Challenges.
          items:
            $ref: '#/components/schemas/ChallengeRequest'
      required:
        - challenges
...
    ReferenceData:
      type: object
      description: A reference item belonging to a particular domain
      properties:
        code:
          type: string
          description: The short code of a reference data
        description:
          type: string
          description: The description of the reference data code
        categoryCode:
          type: string
          description: The short code of a reference data category
        categoryDescription:
          type: string
          description: The description of the reference data category
        areaCode:
          type: string
          description: The short code of a reference data area
        areaDescription:
          type: string
          description: The description of the reference data area
        listSequence:
          type: integer
          format: int32
          description: 'The sequence number of the code. Used for ordering codes correctly in lists and drop downs. '
          example: 3
        active:
          type: boolean
          description: whether or not this ref data is active
      required:
        - code
...
    ChallengeResponse:
      title: ChallengeResponse
      type: object
      allOf:
        - $ref: '#/components/schemas/AuditedResource'
      description: |
        A challenge a person has.
      properties:
        reference:
          type: string
          format: uuid
          description: The unique reference of this challenge
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        fromALNScreener:
          type: boolean
          description: |
            Whether or not this challenge is from the ALN screener
        alnScreenerDate:
          type: string
          format: date
          description: The date that the screener was recorded that has this challenge
          example: '2023-11-19'
        symptoms:
          type: string
          description: Optional details about the symptoms of this challenge
          example: John struggles to read text on white background.
        howIdentified:
          type: array
          items:
            $ref: '#/components/schemas/IdentificationSource'
          description: |
            The list of enums as to how this challenge was identified. Multiple reasons may apply.
        howIdentifiedOther:
          type: string
          description: Details about how this challenge was identified when the OTHER IdentificationSource enum option is chosen
          example: The trainer noticed that John could read better on a cream background.
        challengeType:
          $ref: '#/components/schemas/ReferenceData'
        active:
          type: boolean
          description: |
            Whether or not this challenge is active
      required:
        - reference
        - fromALNScreener
        - challengeType
        - active
...
    ChallengeRequest:
      title: ChallengeRequest
      type: object
      description: |
        A challenge a person has.
      properties:
        challengeTypeCode:
          type: string
          description: the reference data type code of the challenge
        prisonId:
          $ref: '#/components/schemas/PrisonId'
        symptoms:
          type: string
          description: Optional details about the symptoms of this challenge
          example: John struggles to read text on white background.
        howIdentified:
          type: array
          items:
            $ref: '#/components/schemas/IdentificationSource'
          description: |
            The list of enums as to how this challenge was identified. Multiple reasons may apply.
        howIdentifiedOther:
          type: string
          description: Details about how this challenge was identified when the OTHER IdentificationSource enum option is chosen
          example: The trainer noticed that John could read better on a cream background.
      required:
        - challengeTypeCode
        - prisonId
    UpdateChallengeRequest:
      title: UpdateChallengeRequest
      type: object
      description: |
        request object to update a challenge.
      properties:
        symptoms:
          type: string
          description: Details about the symptoms of this challenge
          example: John struggles to read text on white background.
        howIdentified:
          type: array
          items:
            $ref: '#/components/schemas/IdentificationSource'
          description: |
            The list of enums as to how this challenge was identified. Multiple reasons may apply.
        howIdentifiedOther:
          type: string
          description: Details about how this challenge was identified when the OTHER IdentificationSource enum option is chosen
          example: The trainer noticed that John could read better on a cream background.
        prisonId:
          $ref: '#/components/schemas/PrisonId'
          description: The prison the challenge was updated at
      required:
        - symptoms
        - howIdentified
        - prisonId
...
    IdentificationSource:
      title: IdentificationSource
      description: How a challenge was identified.
      type: string
      enum:
        - EDUCATION_SKILLS_WORK
        - WIDER_PRISON
        - CONVERSATIONS
        - COLLEAGUE_INFO
        - FORMAL_PROCESSES
        - SELF_DISCLOSURE
        - OTHER_SCREENING_TOOL
        - OTHER
...
    CreateSupportStrategiesRequest:
      title: SupportStrategiesListRequest
      type: object
      properties:
        supportStrategies:
          type: array
          items:
            $ref: '#/components/schemas/SupportStrategyRequest'
      required:
        - supportStrategies
    UpdateSupportStrategyRequest:
      title: UpdateSupportStrategyRequest
      type: object
      properties:
        detail:
          type: string
          description: Optional further explanation or description of this strategy
          minLength: 1
          pattern: '.*\S.*'
        prisonId:
          $ref: '#/components/schemas/PrisonId'
      required:
        - detail
        - prisonId
    SupportStrategyRequest:
      title: SupportStrategyRequest
      type: object
      properties:
        supportStrategyTypeCode:
          type: string
        prisonId:
          $ref: '#/components/schemas/PrisonId'
        detail:
          type: string
          description: Optional further explanation or description of this strategy
      required:
        - supportStrategyTypeCode
        - prisonId
    SupportStrategyResponse:
      title: SupportStrategyResponse
      allOf:
        - $ref: '#/components/schemas/AuditedResource'
      properties:
        reference:
          type: string
          format: uuid
        detail:
          type: string
        supportStrategyType:
          $ref: '#/components/schemas/ReferenceData'
        active:
          type: boolean
      required:
        - reference
        - supportStrategyType
        - active
    SupportStrategyListResponse:
      title: SupportStrategyListResponse
      type: object
      properties:
        supportStrategies:
          type: array
          items:
            $ref: '#/components/schemas/SupportStrategyResponse'
      required:
        - supportStrategies


  #
  # Response Body Definitions
  # --------------------------------------------------------------------------------
  responses:
    400Error:
      description: Error response for a HTTP 400 describing fields in the request payload that are in error.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            userMessage: 'Error on field deadlineDate: rejected value [2023-06-01] cannot be in the past'
    404Error:
      description: |
        Error response for a HTTP 404 describing the problem with the request.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            userMessage: 'The prisoner with prisonNumber=[A1234BC] not found'
    409Error:
      description: |
        Error response for a HTTP 409 describing the problem with the request.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 409
            userMessage: 'A resource with reference [c88a6c48-97e2-4c04-93b5-98619966447b] already exists for prisoner [A1234BC]'
...
    ChallengeList:
      description: Response body containing the a list of challenges.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChallengeListResponse'
...
    Challenge:
      description: Response body containing a single challenge.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChallengeResponse'
...
    SupportStrategy:
      description: Response body containing a single support strategy.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SupportStrategyResponse'
    SupportStrategyList:
      description: Response body containing the list of support strategies.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SupportStrategyListResponse'
...
  parameters:
...
    prisonNumberPathParameter:
      name: prisonNumber
      description: The NOMIS Prison Reference Number of the Prisoner that the operation is being performed on.
      schema:
        type: string
        pattern: "^[A-Z]\\d{4}[A-Z]{2}$"
        example: 'A1234BC'
      in: path
      required: true
...